+++ packages/office-ui-fabric-react/src/components/CommandBar/CommandBar.tsx
@@ -64,9 +64,8 @@ export class CommandBar extends BaseComponent<ICommandBarProps, ICommandBarState
   }
 
   public componentDidMount() {
-    this._updateItemMeasurements();
-    this._updateRenderedItems();
-
+    // Asynchronously update command bar layout to eliminate forced synchronous reflow
+    this._asyncMeasure();
     this._events.on(window, 'resize', this._updateRenderedItems);
   }
 
@@ -77,8 +76,8 @@ export class CommandBar extends BaseComponent<ICommandBarProps, ICommandBarState
 
   public componentDidUpdate(prevProps: ICommandBarProps, prevStates: ICommandBarState) {
     if (!this._commandItemWidths) {
-      this._updateItemMeasurements();
-      this._updateRenderedItems();
+      // Asynchronously update command bar layout to eliminate forced synchronous reflow
+      this._asyncMeasure();
     }
   }
 
@@ -94,7 +93,7 @@ export class CommandBar extends BaseComponent<ICommandBarProps, ICommandBarState
           <div className={ css(
             'ms-CommandBarSearch-iconWrapper ms-CommandBarSearch-iconSearchWrapper',
             styles.searchIconWrapper, styles.searchIconSearchWrapper) }>
-            <Icon iconName='search' />
+            { Icon({ iconName: 'Search' }) }
           </div>
           <div className={ css(
             'ms-CommandBarSearch-iconWrapper ms-CommandBarSearch-iconClearWrapper',
@@ -111,7 +110,7 @@ export class CommandBar extends BaseComponent<ICommandBarProps, ICommandBarState
     return (
       <div className={ css('ms-CommandBar', styles.root, className) } ref='commandBarRegion'>
         { searchBox }
-        <FocusZone ref='focusZone' direction={ FocusZoneDirection.horizontal } role='menubar' >
+        <FocusZone ref='focusZone' className={ styles.container } direction={ FocusZoneDirection.horizontal } role='menubar' >
           <div className={ css('ms-CommandBar-primaryCommands', styles.primaryCommands) } ref='commandSurface'>
             { renderedItems.map((item, index) => (
               this._renderItemInCommandBar(item, index, expandedMenuItemKey)
@@ -150,8 +149,9 @@ export class CommandBar extends BaseComponent<ICommandBarProps, ICommandBarState
             labelElementId={ expandedMenuId }
             onDismiss={ this._onContextMenuDismiss }
           />
-          ) : (null) }
-      </div>
+          ) : (null)
+        }
+      </div >
     );
   }
 
@@ -197,7 +197,7 @@ export class CommandBar extends BaseComponent<ICommandBarProps, ICommandBarState
               </span>
             ) }
             { hasSubmenuItems(item) ? (
-              <Icon className={ css('ms-CommandBarItem-chevronDown', styles.itemChevronDown) } iconName='chevronDown' />
+              <Icon className={ css('ms-CommandBarItem-chevronDown', styles.itemChevronDown) } iconName='ChevronDown' />
             ) : (null) }
           </button>;
         } else if (item.href) {
@@ -247,6 +247,13 @@ export class CommandBar extends BaseComponent<ICommandBarProps, ICommandBarState
     let iconClassName = css('ms-CommandBarItem-icon', styles.itemIcon, iconColorClassName, iconProps.className);
 
     return <Icon { ...iconProps } className={ iconClassName } />;
+  }
+
+  private _asyncMeasure() {
+    this._async.requestAnimationFrame(() => {
+      this._updateItemMeasurements();
+      this._updateRenderedItems();
+    });
   }
 
   private _updateItemMeasurements() {