+++ packages/styling/src/styles/DefaultFontStyles.ts
@@ -118,7 +118,8 @@ function _getFontFamily(): string {
 
   for (let lang in LanguageToFontMap) {
     if (LanguageToFontMap.hasOwnProperty(lang) && lang.indexOf(language) === 0) {
-      fontFamily = LanguageToFontMap[lang];
+      // tslint:disable-next-line:no-any
+      fontFamily = (LanguageToFontMap as any)[lang];
       break;
     }
   }
@@ -204,8 +205,8 @@ function _registerDefaultFontFaces(): void {
 function _getFontBaseUrl(): string {
   let win = getWindow();
 
-  // tslint:disable-next-line:no-string-literal
-  let fabricConfig: IFabricConfig = win ? win['FabricConfig'] : undefined;
+  // tslint:disable-next-line:no-string-literal no-any
+  let fabricConfig: IFabricConfig = win ? (win as any)['FabricConfig'] : undefined;
 
   return (fabricConfig && fabricConfig.fontBaseUrl !== undefined) ? fabricConfig.fontBaseUrl : DefaultBaseUrl;
 }