+++ apps/fabric-website/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log - @uifabric/fabric-website
 
-This log was last generated on Wed, 14 Jun 2017 06:02:15 GMT and should not be manually modified.
+This log was last generated on Tue, 27 Jun 2017 01:26:31 GMT and should not be manually modified.
+
+## 4.5.40
+Tue, 27 Jun 2017 01:26:31 GMT
+
+### Patches
+
+- Enable forceConsistentCasingInFileNames tsconfig option
+
+## 4.5.39
+Wed, 21 Jun 2017 00:45:41 GMT
+
+*Changes not tracked*
 
 ## 4.5.38
 Wed, 14 Jun 2017 06:02:15 GMT